+++ package.json
@@ -1,0 +1,91 @@
+{
+  "name": "jspro2",
+  "version": "1.0.0",
+  "description": "",
+  "main": "script.js",
+  "dependencies": {
+    "accepts": "^1.3.8",
+    "body-parser": "^1.20.0",
+    "bytes": "^3.1.2",
+    "array-flatten": "^1.1.1",
+    "call-bind": "^1.0.2",
+    "content-disposition": "^0.5.4",
+    "content-type": "^1.0.4",
+    "cookie": "^0.5.0",
+    "cookie-signature": "^1.0.6",
+    "cors": "^2.8.5",
+    "debug": "^2.6.9",
+    "depd": "^2.0.0",
+    "destroy": "^1.2.0",
+    "ee-first": "^1.1.1",
+    "encodeurl": "^1.0.2",
+    "escape-html": "^1.0.3",
+    "etag": "^1.8.1",
+    "express": "^4.18.1",
+    "finalhandler": "^1.2.0",
+    "forwarded": "^0.2.0",
+    "fresh": "^0.5.2",
+    "function-bind": "^1.1.1",
+    "get-intrinsic": "^1.1.2",
+    "has": "^1.0.3",
+    "has-symbols": "^1.0.3",
+    "http-errors": "^2.0.0",
+    "iconv-lite": "^0.4.24",
+    "inherits": "^2.0.4",
+    "ipaddr.js": "^1.9.1",
+    "media-typer": "^0.3.0",
+    "merge-descriptors": "^1.0.1",
+    "methods": "^1.1.2",
+    "mime": "^1.6.0",
+    "mime-db": "^1.52.0",
+    "mime-types": "^2.1.35",
+    "ms": "^2.0.0",
+    "negotiator": "^0.6.3",
+    "object-assign": "^4.1.1",
+    "on-finished": "^2.4.1",
+    "object-inspect": "^1.12.2",
+    "parseurl": "^1.3.3",
+    "path-to-regexp": "^0.1.7",
+    "proxy-addr": "^2.0.7",
+    "qs": "^6.10.3",
+    "range-parser": "^1.2.1",
+    "raw-body": "^2.5.1",
+    "safe-buffer": "^5.2.1",
+    "safer-buffer": "^2.1.2",
+    "send": "^0.18.0",
+    "serve-static": "^1.15.0",
+    "setprototypeof": "^1.2.0",
+    "side-channel": "^1.0.4",
+    "statuses": "^2.0.1",
+    "toidentifier": "^1.0.1",
+    "type-is": "^1.6.18",
+    "unpipe": "^1.0.0",
+    "utils-merge": "^1.0.1",
+    "vary": "^1.1.2"
+  },
+  "devDependencies": {
+    "css-loader": "^6.7.1",
+    "html-webpack-plugin": "^5.5.0",
+    "jest": "^28.1.0",
+    "style-loader": "^3.3.1",
+    "webpack": "^5.72.1",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.9.0"},
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "webpack serve --open",
+    "build": "webpack"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ZenkovaLyubov/JSPRO.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "type": "module",
+  "bugs": {
+    "url": "https://github.com/ZenkovaLyubov/JSPRO/issues"
+  },
+  "homepage": "https://github.com/ZenkovaLyubov/JSPRO#readme"
+}
