+++ src/components/CustomButton/index.js
@@ -1,0 +1,8 @@
+export default Vue.component('custom-button', {
+    template:
+    `
+    <button class="button cart-button" type="button" @click="$emit('click')">
+    <slot></slot>
+    </button>
+    `
+  })
