+++ src/constants.js
@@ -1,0 +1,4 @@
+export const BASE_URL = "http://localhost:8000/";
+export const GOODS_ADD = `${BASE_URL}goods`;
+export const GOODS = `${BASE_URL}goods.json`;
+export const GOODS_BASKET = `${BASE_URL}basket`;
