+++ src/index.html
@@ -5,10 +5,9 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>eShop</title>
-    <link rel="stylesheet" href="style.css">
     <script src="https://cdn.jsdelivr.net/npm/vue@2.6.14/dist/vue.js"></script>
     <script src="https://kit.fontawesome.com/04922efd3c.js" crossorigin="anonymous"></script>
-    <script src="script.js"></script>
+
 </head>
 
 <body>
